add ephemeral example resource (#264)
* eph: add ephemeral resource

Adds a new example ephemeral resource using a required configurable attribute and an example ID.

* tools: bump terraform-plugin-docs

Bump to the latest stable release, as it includes support for ephemeral resource documentation generation.

* docs: add ephemeral resource documentation

* provider: register new example ephemeral resource

* eph: improve alphabetical ordering of struct fields

* eph: use a computed 'value' attribute

To help provider developers easily understand the concept of ephemeral resources, we use a 'value' attribute in the schema, which is set to an example token (token-123).

* docs: update ephemeral resource docs

* provider: create a separate echo provider factory container

Since not every acceptance test requires an ephemeral provider server for the CLI to connect to and interact with, this introduces a separate provider factory container that includes the echo provider.

* eph: add basic acceptance test for ephemeral resource

* docs: rename ephemeral resource configuration file

* eph: improve http response diag